backend/requirements.txt
# FastAPI and core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0
python-multipart==0.0.12

# Firebase
firebase-admin==6.3.0

# Database
motor==3.3.2  # Async MongoDB driver
pymongo==4.6.0

# External API clients
# No paid APIs required - using 100% free sources!

# ML dependencies
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.1.3
joblib==1.3.2

# Networking and enrichment
dnspython==2.4.2
httpx==0.25.0
aiohttp==3.9.1

# Stripe
stripe==7.8.0

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Utilities
python-dotenv==1.0.0

# Development and testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.0  # For testing
